requirements.txt
groq==0.4.1
python-dotenv